Update Get_ContamReproductionEffects in atContaminants.c line 1379
original:
if (cGroupLevel > chronicLevel) {
        step1 = bm->contaminantStructure[cIndex]->sp_ReprodEffect[species] * cGroupLevel;
        if (step1 > 1.0)
            step1 = 1.0;
        step2 = 1.0 - step1;

chronicLevel was used in the if statement but not used within the equation, and concentrations above chronicLevel were resuting in step2 = 0.
The sp_ReprodEffect was not getting used.

updated equation to:
if (cGroupLevel > chronicLevel) {
        step1 = bm->contaminantStructure[cIndex]->sp_ReprodEffect[species] * (cGroupLevel -  chronicLevel);
        if (step1 > 1.0)
            step1 = 1.0;
        step2 = 1.0 - step1;

Doing simulations with high, low and no contaminant uptake for herring to see impacts and understand contaminant transfer.